# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Sup Seasonal
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /foods:
    x-swagger-router-controller: foods
    get:
      consumes:
        - application/json
      summary: Gets all foods
      description: Returns a list of all foods and their seasonal information.
      operationId: getAllFoods
      responses:
        "200":
          description: Foods response
          schema:
            $ref: "#/definitions/FoodsList"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      consumes:
        - application/json
      summary: Adds a food
      description: Adds a new food and its seasonal information to the foods list.
      operationId: addFood
      parameters:
        - name: food
          in: body
          description: The food to add.
          required: true
          schema:
            $ref: "#/definitions/Food"
      responses:
        "200":
          description: Added food response
          schema:
            $ref: "#/definitions/Food"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /foods/{food_id}:
    x-swagger-router-controller: foods
    get:
      summary: Gets a single food item
      description: Returns all seasonal information about a given food.
      operationId: getFood
      parameters:
        - name: food_id
          in: path
          description: ID of food to get
          required: true
          type: integer
      responses:
        "200":
          description: Food response
          schema:
            $ref: "#/definitions/Food"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      summary: Update a single food item
      description: Updates the seasonal information about a given food item.
      operationId: updateFood
      parameters:
        - name: food_id
          in: path
          description: ID of food to update
          required: true
          type: integer
        - name: food
          in: body
          required: true
          schema:
            $ref: "#/definitions/Food"
      responses:
        "200":
          description: Updated food response
          schema:
            $ref: "#/definitions/Food"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      summary: Removes a food
      description: Removes a food from the foods list.
      operationId: deleteFood
      parameters:
        - name: food_id
          in: path
          description: ID of food to delete
          required: true
          type: integer
      responses:
        "204":
          description: Food deleted
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /months/{month}:
    x-swagger-router-controller: months
    get:
      summary: Gets all foods for a given month
      description: Returns a list of all foods and their seasonal information for a specific month.
      operationId: getMonth
      parameters:
        - name: month
          in: path
          description: Month of the year
          required: true
          type: string
      responses:
        "200":
          description: Foods response
          schema:
            $ref: "#/definitions/FoodsList"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /months/{month}/recipes:
    x-swagger-router-controller: months
    get:
      summary: Gets recipes using seasonal ingredients for a given month
      description: Returns a list of recipes and their seasonal information for a specific month.
      operationId: getRecipes
      parameters:
        - name: month
          in: path
          description: Month of the year
          required: true
          type: string
      responses:
        "200":
          description: Recipes response
          schema:
            $ref: "#/definitions/RecipeList"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /months/{month}/favorites:
    x-swagger-router-controller: favorites
    get:
      summary: Gets favorited recipes by month
      description: Returns a list of all favorited recipes for ingredients from a specific month.
      operationId: getFavoritesByMonth
      parameters:
        - name: month
          in: path
          description: Month of the year
          required: true
          type: string
      responses:
        "200":
          description: Favorites by month response
          schema:
            $ref: "#/definitions/FavoritesByMonthList"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users:
    x-swagger-router-controller: users
    get:
      summary: Gets all users
      description: Returns a list of all users and their information.
      operationId: getAllUsers
      responses:
        "200":
          description: Users response
          schema:
            $ref: "#/definitions/UsersList"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      summary: Add a user
      description: Adds a new user to the users list.
      operationId: addUser
      parameters:
        - name: user
          in: body
          description: The user to add.
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Added user response
          schema:
            $ref: "#/definitions/User"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{user_id}:
    x-swagger-router-controller: users
    get:
      summary: Gets a user
      description: Returns a single user.
      produces:
       - application/json
      operationId: getUser
      parameters:
        - name: user_id
          in: path
          description: User's ID
          required: true
          type: integer
      responses:
        "200":
          description: User response
          schema:
            $ref: "#/definitions/User"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      summary: Update a single user
      description: Updates a user's information.
      operationId: updateUser
      parameters:
        - name: user_id
          in: path
          description: User's ID
          required: true
          type: integer
        - name: user
          in: body
          description: The user to update.
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Updated user response
          schema:
            $ref: "#/definitions/User"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      summary: Removes a user
      description: Removes a user from the users list.
      operationId: deleteUser
      parameters:
        - name: user_id
          in: path
          description: User's ID
          required: true
          type: integer
      responses:
        "204":
          description: User deleted
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /favorites:
    x-swagger-router-controller: favorites
    get:
      summary: Gets all favorited recipes
      description: Returns a list of all of a user's favorited recipes.
      operationId: getFavorites
      responses:
        "200":
          description: Favorites response
          schema:
            $ref: "#/definitions/FavoritesList"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /favorites/{recipe_id}:
    x-swagger-router-controller: favorites
    get:
      summary: Gets favorited recipe
      description: Returns a single recipe saved to a users favorites.
      operationId: getFavRecipeById
      parameters:
        - name: recipe_id
          in: path
          description: ID of favorited recipe
          required: true
          type: integer
      responses:
        "200":
          description: Favorite recipe by ID response
          schema:
            $ref: "#/definitions/Favorite"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  FoodsList:
    required:
      - foods
    properties:
      foods:
        type: array
        items:
          $ref: "#/definitions/Food"
  Food:
    required:
      - food_name
    properties:
      food_name:
        type: string
  RecipeList:
    required:
      - recipes
    properties:
      recipes:
        type: array
        items:
          $ref: "#/definitions/Recipe"
  Recipe:
    required:
      - title
      - image
      - imageType
      - usedIngredientCount
      - missedIngredientCount
      - likes
    properties:
      title:
        type: string
      image:
        type: string
      imageType:
        type: string
      usedIngredientCount:
        type: integer
      missedIngredientCount:
        type: integer
      likes:
        type: integer
  FavoritesList:
    required:
      - favorites
    properties:
      favorites:
        type: array
        items:
          $ref: "#/definitions/Favorite"
  FavoritesByMonthList:
    required:
      - favorites
    properties:
      favorites:
        type: array
        items:
          $ref: "#/definitions/Favorite"
  Favorite:
    required:
      - user_id
      - recipe_id
      - month
    properties:
      user_id:
        type: integer
      recipe_id:
        type: integer
      month:
        type: string
  UsersList:
    required:
      - users
    properties:
      users:
        type: array
        items:
          $ref: "#/definitions/User"
  User:
    required:
      - username
      - email
      - permissions
      - hashed_password
    properties:
      username:
        type: string
      email:
        type: string
      permissions:
        type: string
        default: 'user'
      hashed_password:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string